pyproject.toml
[project]
name = "lung-detect"
version = "0.1.0"
description = "Lung Cancer Detection Application using AI"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "keras>=3.0.0",
    "matplotlib>=3.10.8",
    "numpy>=2.2.6",
    "opencv-python>=4.12.0.88",
    "pillow>=10.0.0",
    "pyside6>=6.6.0",
    "scikit-learn>=1.8.0",
    "seaborn>=0.13.2",
    "tensorflow[and-cuda]>=2.20.0",
]

[project.scripts]
lung-detect = "src.main:main"

[dependency-groups]
dev = ["pytest>=8.0.0", "pytest-qt>=4.2.0"]
